Change response structure

+++ app/crud/catalogs/currency_crud.py
@@ -53,15 +53,18 @@ def update_currency_by_id(db, currency_id, modify_currency):
 
 def delete_currency(db: Session, currency_id:int):
     currency = db.query(Currency).filter(Currency.id==currency_id).first()
+    if currency == None:
+        return None
     try:
         db.delete(currency)
         db.commit()
     except IntegrityError:
         db.rollback()
-        return {"status": 2, "detail": "Can´t delete currency, currency referenced by other table"}    
-    except:
+        return {"status": 2, "msg": "Can´t delete currency, currency referenced by other table"}    
+    except Exception as e:
+        print(e)
         db.rollback()
-        return {"status": 3, "detail": "Internal Server Error"}
+        return {"status": 3, "msg": "Internal Server Error"}
     
     return {"status" : 1,
-            "detail": "Currency deleted successfully"}
+            "msg": "Currency deleted successfully"}
